Apply VSG_DECLSPEC to Array2D, Array3D, and Value; move VSG_USE_DYNAMIC_CAST to version.h.in

+++ include/vsg/core/Array.h
@@ -101,7 +101,7 @@ namespace vsg
             _data(nullptr),
             _size(0)
         {
-            assign(data, offset, stride, static_cast<uint32_t>(l.size()));
+            assign(data, offset, stride, l.size());
 
             iterator itr = begin();
             for (const value_type& v : l) { (*itr++) = v; }
